+++ src/commonMain/kotlin/io/github/koalaplot/core/xygraph/FloatLinearAxisModel.kt
@@ -45,7 +45,7 @@ public class FloatLinearAxisModel(
     private val allowZooming: Boolean = true,
     private val allowPanning: Boolean = true,
     private val inverted: Boolean = false,
-) : AxisModel<Float>, LinearAxisModel<Float> {
+) : ContinuousLinearAxisModel<Float> {
     init {
         require(range.endInclusive > range.start) {
             "Axis range end (${range.endInclusive}) must be greater than start (${range.start})"