+++ src/commonMain/kotlin/io/github/koalaplot/core/xygraph/LinearAxisModel.kt
@@ -8,10 +8,25 @@ internal val TickRatios = listOf(0.1f, 0.2f, 0.5f, 1f, 2f)
 internal const val ZoomRangeLimitDefault = 0.2
 internal const val MinimumMajorTickIncrementDefault = 0.1f
 
+/**
+ * Interface implemented by all [AxisModel]s that transform linearly from plot coordinates to screen coordinates.
+ */
 public interface LinearAxisModel<T> : AxisModel<T> where T : Comparable<T>, T : Number {
+    /**The minimum to maximum values allowed to be represented on this Axis. Zoom and
+     * scroll modifications may not exceed this range.
+     * **/
     public val range: ClosedRange<T>
-    public val minimumMajorTickSpacing: Dp
 
+    /**
+     * Specifies the minimum physical spacing for major ticks, in Dp units. Must be greater than 0.
+     */
+    public val minimumMajorTickSpacing: Dp
+}
+
+/**
+ * Interface implemented by [LinearAxisModel]s that use real numbers.
+ */
+public interface ContinuousLinearAxisModel<T> : LinearAxisModel<T> where T : Comparable<T>, T : Number {
     public val viewRange: State<ClosedRange<T>>
 
     /**
@@ -20,6 +35,23 @@ public interface LinearAxisModel<T> : AxisModel<T> where T : Comparable<T>, T : 
      * [range] of the axis. If they exceed either limit they will be clamped to the range.
      */
     public fun setViewRange(newRange: ClosedRange<T>)
+}
+
+/**
+ * Interface implemented by [LinearAxisModel]s that use integers.
+ */
+public interface DiscreteLinearAxisModel<T> : LinearAxisModel<T> where T : Comparable<T>, T : Number {
+    /**
+     * The currently viewable range on the axis.
+     */
+    public val viewRange: State<ClosedRange<Double>>
+
+    /**
+     * Sets the currently viewable range on the axis. This is a more direct alternative to using [pan] and [zoom]
+     * to set the range of currently viewable values on an axis. The start and end must be within the allowed
+     * [range] of the axis. If they exceed either limit they will be clamped to the range.
+     */
+    public fun setViewRange(newRange: ClosedRange<Double>)
 }
 
 internal fun <X, Y> List<Point<X, Y>>.toXList(): List<X> = object : AbstractList<X>() {